// client/src/entities/consts/user_names.ts
export enum user_names {
  whorm = 'Червь',
  baron = 'Барон Харконнен',
  mother = 'Преподобная Мать',
  duke = 'Герцог Лето',
  paul = 'Пол Атрейдес',
  jessica = 'Леди Джессика',
  stilgar = 'Стилгар',
  chani = 'Чани',
  idaho = 'Дункан Айдахо',
  halec = 'Гурни Халлек',
  covid = 'Covid-19',
}